pyproject.toml
[project]
name = "pyharborcli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexeev Bronislav", email = "alexeev.dev@mail.ru" }
]
requires-python = ">=3.10"
dependencies = [
    "aiodocker>=0.24.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "asyncclick>=8.1.8",
    "asyncio>=3.4.3",
    "orjson>=3.10.18",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "toml>=0.10.2",
]

[project.scripts]
pyharborcli = "pyharborcli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mutmut>=3.3.0",
    "mypy>=1.16.1",
    "nox>=2025.5.1",
    "pytest>=8.4.1",
    "pytest-coverage>=0.0",
    "pytest-randomly>=3.16.0",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.2",
]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  "A",    # flake8-builtins
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "C90",  # maccabe
  "COM",  # flake8-commas
  "D",    # pydocstyle
  "DTZ",  # flake8-datetimez
  "E",    # pycodestyle
  "ERA",  # flake8-eradicate
  "EXE",  # flake8-executable
  "F",    # pyflakes
  "FBT",  # flake8-boolean-trap
  "FLY",  # pyflint
  "FURB", # refurb
  "G",    # flake8-logging-format
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PL",   # pylint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "Q",    # flake8-quotes
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # ruff
  "S",    # flake8-bandit
  "SIM",  # flake8-simpify
  "SLF",  # flake8-self
  "SLOT", # flake8-slots
  "T100", # flake8-debugger
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
ignore = [
  "A005",   # allow to shadow stdlib and builtin module names
  "COM812", # trailing comma, conflicts with `ruff format`
  # Different doc rules that we don't really care about:
  "D100",
  "D104",
  "D106",
  "D203",
  "D212",
  "D401",
  "D404",
  "D405",
  "ISC001",  # implicit string concat conflicts with `ruff format`
  "ISC003",  # prefer explicit string concat over implicit concat
  "PLR09",   # we have our own complexity rules
  "PLR2004", # do not report magic numbers
  "PLR6301", # do not require classmethod / staticmethod when self not used
  "TRY003",  # long exception messages from `tryceratops`
]
